Add to Workspace ajax types.
